HARmony_Backend/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
tensorflow==2.20.0
numpy>=1.26.0
pandas>=2.2.0
reportlab==4.0.9
psycopg2-binary==2.9.10
